- allow both adding existing images and creating new image options - make the dialog pretty - ask the user whether they want to delete the image or not
git-svn-id: file:///srv/svn/reactos/trunk@27320 97493ccd-5924-5043-b1f5-66cb403b36ce

+++ tools/RosTE/GUI/SettingsForm.cs
@@ -352,6 +352,21 @@ namespace RosTEGUI
         {
             int oldSel = harddiskLstBox.SelectedIndex;
             VMHardDrive vmhd = (VMHardDrive)harddiskLstBox.SelectedItem;
+
+            DialogResult ret;
+            ret = MessageBox.Show("Do you want to delete the image file too?",
+                                  "Deleting harddisk",
+                                  MessageBoxButtons.YesNoCancel,
+                                  MessageBoxIcon.Warning);
+            if (ret == DialogResult.Cancel)
+            {
+                return;
+            }
+            else if (ret == DialogResult.Yes)
+            {
+                ;//FIXME: delete the image
+            }
+
             VirtMach.DeleteHardDisk(vmhd);
             harddiskLstBox.Items.Remove(vmhd);
             harddiskLstBox.SelectedIndex = oldSel - 1;
